+++ R/base.R
@@ -89,3 +89,14 @@ unique.ypr_populations <- function(x, ...) {
   as_ypr_populations(x)
 }
 
+#' @describeIn ypr_population_update Update Population Parameters
+#' @export
+update.ypr_population <- function(object, ...) {
+  ypr_population_update(object, ...)
+}
+
+#' @describeIn ypr_populations_update Update Populations Parameters
+#' @export
+update.ypr_populations <- function(object, ...) {
+  ypr_population_update(object, ...)
+}