+++ R/namespace.R
@@ -1,5 +1,5 @@
 #' @import ggplot2 chk yesno
 #' @importFrom purrr map map2 transpose
-#' @importFrom stats setNames
+#' @importFrom stats update setNames
 #' @importFrom graphics plot
 NULL